pkcs11: explicit conversions with loss of precision in cases that do not matter
(cherry picked from commit 710cfbcfbbef806fef49aee71f15beb1a95d0c4a)

+++ cardcomm/pkcs11/src/cal.cpp
@@ -219,7 +219,7 @@ CK_RV cal_init_slots(void)
 
 	try
 	{
-		nReaders = oReadersInfo->ReaderCount();
+		nReaders = (unsigned int)(oReadersInfo->ReaderCount());
 		//get readernames
 		for (i = 0; i < nReaders; i++)
 		{
